fix: moved beforeunload logic to surveymenubar
